pndevmapperd: add unexpected charge stop workaround
My machine in some real rare cases (once a week or so) just stops
charging without any good reason. What helps here is replugging
the charger or writing to a chip charge register, so we do the later
indirectly every 20 minutes as a workaround.

There have been similar reports in the forums, where user leaves
the machine charging overnight and finds flat battery in the morning,
so hopefully this will help them too.